refactor: remove pkg folder from cr discovery server
