+++ containerruntimediscovery/server/cmd/main.go
@@ -27,7 +27,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	server "github.com/openclarity/vmclarity/containerruntimediscovery/server/pkg"
+	server "github.com/openclarity/vmclarity/containerruntimediscovery/server"
 	"github.com/openclarity/vmclarity/core/log"
 )
